requirements.txt
line-bot-sdk
flask
configparser
beautifulsoup4
requests